requirements.txt
selenium
webdriver-manager
python-telegram-bot
